mgenerate-core/src/main/java/uk/dioxic/mgenerate/core/operator/chrono/Second.java
package uk.dioxic.mgenerate.core.operator.chrono;

import uk.dioxic.mgenerate.common.Resolvable;
import uk.dioxic.mgenerate.common.annotation.Operator;
import uk.dioxic.mgenerate.common.annotation.OperatorProperty;
import uk.dioxic.mgenerate.core.operator.AbstractOperator;
import uk.dioxic.mgenerate.core.util.FakerUtil;

import java.time.LocalDateTime;

@Operator
public class Second extends AbstractOperator<Integer> {

    @OperatorProperty
    Resolvable<LocalDateTime> date;

    @Override
    public Integer resolveInternal() {
        return (date != null) ? date.resolve().getSecond() : FakerUtil.numberBetween(0, 60);
    }

}
